requirements.txt
alembic==1.14.0
apispec==6.8.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
loginform==1.2.0
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.25.1
packaging==24.2
python-dotenv==1.0.1
PyYAML==6.0.2
SQLAlchemy==2.0.37
typing_extensions==4.12.2
webargs==8.6.0
Werkzeug==3.1.3
WTForms==3.2.1
